Crate uchardet

A wrapper around the uchardet library. Detects character encodings.

Note that the underlying implemention is written in C and C++, and I'm not aware of any security audits which have been performed against it.

use uchardet::detect_encoding_name;

assert_eq!(Some("windows-1252".to_string()),
           detect_encoding_name(&[0x66u8, 0x72, 0x61, 0x6e, 0xe7,
                                0x61, 0x69, 0x73]).unwrap());

For more information, see this project on GitHub.

Structs

EncodingDetectorError
An error occurred while trying to detect the character encoding.

Functions

detect_encoding_name
Return the name of the charset used in data, or None if the charset is ASCII or if the encoding can't be detected. This is the value returned by the underlying uchardet library, with the empty string mapped to None.

Type Definitions

EncodingDetectorResult
Either a return value, or an encoding detection error.