Function uchardet::detect_encoding_name

pub fn detect_encoding_name(data: &[u8]) -> EncodingDetectorResult<Option<String>>

Return the name of the charset used in data, or None if the charset is ASCII or if the encoding can't be detected. This is the value returned by the underlying uchardet library, with the empty string mapped to None.

use uchardet::detect_encoding_name;

assert_eq!(None, detect_encoding_name("ascii".as_bytes()).unwrap());
assert_eq!(Some("UTF-8".to_string()),
           detect_encoding_name("français".as_bytes()).unwrap());
assert_eq!(Some("windows-1252".to_string()),
           detect_encoding_name(&[0x66u8, 0x72, 0x61, 0x6e, 0xe7,
                                0x61, 0x69, 0x73]).unwrap());